+++ content_upload.go
@@ -120,22 +120,6 @@ func (cuh *ContentUploadHandler) addFiles(files []ContentFile, contentType Conte
 		}
 
 		if contentType == ContentTypeCar {
-			if _, name := filepath.Split(file.FilePath); name == "data.acd" {
-				err := addTyresFromDataACD(file.FilePath, fileDecoded)
-
-				if err != nil {
-					logrus.WithError(err).Errorf("Could not create tyres for new car (%s)", file.FilePath)
-				}
-			} else if name == "tyres.ini" {
-				// it seems some cars don't pack their data into an ACD file, it's just in a folder called 'data'
-				// so we can just grab tyres.ini from there.
-				err := addTyresFromTyresIni(file.FilePath, fileDecoded)
-
-				if err != nil {
-					logrus.WithError(err).Errorf("Could not create tyres for new car (%s)", file.FilePath)
-				}
-			}
-
 			uploadedCars[parts[0]] = true
 		}